Change the behavior of Find (..., No_Element) for non-empty containers
From returning No_Element to raising Constraint_Error, on
Ada.Containers.Vectors and Doubly_Linked_Lists.

The correct behavior according with RM is finding from the first.
However, Find is separated to two overloaded subprograms in drake.
The one does not have Position and finds from the first, and the other
one has Position and finds from it.
This change makes using the latter as the former to be error.